// src/features/booking/dto/booking-shipping-response.dto.ts
export type BookingTransitLegResponseDto = {
  id: number;
  portId: number;
  sortOrder: number;
  eta: Date | null;
  etd: Date | null;
};

export type BookingShippingResponseDto = {
  id?: number;
  bookingPartnerId: number;
  bookingNo: string | null;
  bookingTo: string | null;
  bookingNumberReference: string | null;
  bookingNote: string | null;
  serviceMode: string | null;
  placeOfReceiptPortId: number | null;
  portOfLoadingPortId: number | null;
  pickUp: string | null;
  etd: Date | null;
  dateOfPickUp: string | null;
  dropOffWarehouse: string | null;
  feederVessel: string | null;
  feederVoyage: string | null;
  motherVessel: string | null;
  motherVoyage: string | null;
  provider: string | null;
  carrier: string | null;
  cyCutOff: Date | null;
  siCutOff: Date | null;
  vgmCutOff: Date | null;
  gateIn: Date | null;
  temp: string | null;
  vent: string | null;
  freightTerms: string | null;
  portOfDischargePortId: number | null;
  placeOfDeliveryPortId: number | null;
  finalDestinationPortId: number | null;
  eta: Date | null;
  volume: string | null;
  cargoType: string | null;
  cargoName: string | null;
  grossWeightKgs: string | null;
  measurementCbm: string | null;
  contact: string | null;
  specialRemark: string | null;
  dateOfCreation: string | null;
  termsAndConditions: string | null;
  transitLegs: BookingTransitLegResponseDto[];
};